Create Rules folder in tests and move rule in it
